테스트 환경에서 UniqueInfoContainer 를 단일 스레드에서 더 가볍게 실행하도록 리팩토링. 이전에는 멀티스레드에서 동시성 이슈를 제어하도록 설계된 컨테이너였음
