+++ devlog/240127-1-SpringBootTest.txt
@@ -12,7 +12,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
 import psam.portfolio.sunder.english.global.jpa.embeddable.Address;
 import psam.portfolio.sunder.english.testbean.TestConfig;
-import psam.portfolio.sunder.english.testbean.UniqueInfoContainer;
+import psam.portfolio.sunder.english.testbean.ConcurrentUniqueInfoContainer;
 import psam.portfolio.sunder.english.web.teacher.enumeration.AcademyStatus;
 import psam.portfolio.sunder.english.web.teacher.model.entity.Academy;
 import psam.portfolio.sunder.english.web.teacher.model.entity.Teacher;
